pyproject.toml
[tool.poetry]
name = "TestCapitalComPySe"
version = "3.12.4"
description = ""
authors = ["Alexander Tomelo <qa.tomelo.an@gmail.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = "^3.10.0"
webdriver-manager = "4.0.1"
selenium = "4.22.0"
pytest = "8.2.2"
pytest-html = "^4.0.2"
allure-pytest = "^2.13.5"
pytest-xdist = "^3.5.0"
#pytest-timeout = "^2.2.0"
#pyautogui = "^0.9.53"
#torch = "2.2.2"
psutil = "5.9.8"
#aiogram = "^2.24"
#memory-profiler
speedtest-cli = "^2.1.3"
#qreader = "3.12"
google = "3.0.0"
google-auth = "2.31.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.0"
#google-api-core = "^2.18.0"
google-api-python-client = "2.125.0"
#ruff
flake8 = "^7.0.0"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "ERROR"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_file = "pytest.log"
log_file_level = "ERROR"
log_file_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
